Insert functionality implemented

+++ src/test/java/com/lysenkova/queryapp/queryexecutor/QueryExecutorTest.java
@@ -1,6 +1,7 @@
 package com.lysenkova.queryapp.queryexecutor;
 
 import com.lysenkova.queryapp.entity.Request;
+import com.lysenkova.queryapp.entity.Response;
 import com.lysenkova.queryapp.entity.SQLType;
 import com.lysenkova.queryapp.entity.metadata.Column;
 import org.junit.After;
@@ -13,23 +14,12 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 
 public class QueryExecutorTest {
-    @Before
-    public void before() {
-        Request request = new Request();
-        Map<String, String> header = new HashMap<>();
-        header.put("type", SQLType.valueOf("CREATE").getName());
-        request.setHeader(header);
-        request.setEntity("schema");
-        request.setName("soap");
-        File file = new File("src\\test\\resources\\db\\\\soap");
-        file.mkdir();
-    }
 
     @Test
-    public void createSchema() {
+    public void executeQuerySchemaTest() {
         Request request = new Request();
         Map<String, String> header = new HashMap<>();
         header.put("type", SQLType.valueOf("CREATE").getName());
@@ -38,12 +28,12 @@ public class QueryExecutorTest {
         request.setName("soup");
         QueryExecutor queryExecutor = new QueryExecutor("src\\test\\resources\\db\\");
         queryExecutor.executeQuery(request);
-        File file = new File("src\\test\\resources\\db\\" + request.getName() + ".xml");
-        assertEquals("soup.xml", file.getName());
+        File file = new File("src\\test\\resources\\db\\" + request.getName());
+        assertEquals("soup", file.getName());
     }
 
     @Test
-    public void createTable() {
+    public void executeQueryTableTest() {
         List<Column> columns = new ArrayList<>();
         columns.add(new Column("id", "BIGINT"));
         columns.add(new Column("name", "VARCHAR(255)"));
@@ -62,20 +52,55 @@ public class QueryExecutorTest {
         assertEquals("soap.xml", file.getName());
     }
 
+    @Test
+    public void executeQueryInsertTest() {
+        Request request = new Request();
+        Map<String, String> header = new HashMap<>();
+        header.put("type", "INSERT");
+        header.put("schema", "soap");
+        request.setHeader(header);
+        request.setEntity("table");
+        request.setName("test");
+        List<Column> columns = new ArrayList<>();
+        columns.add(new Column("id", "int", "22"));
+        columns.add(new Column("name", "varchar(255)", "Mark"));
+        request.setColumns(columns);
+        QueryExecutor queryExecutor = new QueryExecutor("src\\test\\resources\\db\\");
+        Response expected = new Response();
+        expected.setMessage("Data inserted successfully.");
+        expected.setEntity("table");
+        expected.setEntityName(request.getName());
+        Response actual = queryExecutor.executeQuery(request);
+        assertEquals(expected, actual);
+    }
+
     @After
     public void clean() {
-        File fileTable = new File("src\\test\\resources\\db\\soap\\");
+        File fileTable = new File("src\\test\\resources\\db\\data\\soap\\");
         File[] filesTable = fileTable.listFiles();
-        for (File file : filesTable) {
-             file.delete();
-            System.out.println(file.getAbsolutePath());
+        if (filesTable != null) {
+            for (File file : filesTable) {
+                file.delete();
+                System.out.println(file.getAbsolutePath());
+            }
         }
-        File file = new File("src\\test\\resources\\db\\");
+        File file = new File("src\\test\\resources\\db\\data\\");
         File[] files = file.listFiles();
         for (File testFile : files) {
             testFile.delete();
         }
-
-
+        File fileTableMeta = new File("src\\test\\resources\\db\\metadata\\soap\\");
+        File[] filesTableMeta = fileTableMeta.listFiles();
+        if (filesTableMeta != null) {
+            for (File fileMeta : filesTableMeta) {
+                fileMeta.delete();
+                System.out.println(fileMeta.getAbsolutePath());
+            }
+        }
+        File fileMeta = new File("src\\test\\resources\\db\\metadata\\");
+        File[] filesMeta = fileMeta.listFiles();
+        for (File testFile : filesMeta) {
+            testFile.delete();
+        }
     }
 }
